added expect module
This is a hygienic rewrite of ice-9 expect macros, which to fix
problems with the original non-higienic defmacro versions.

ice-9 expect has support for procedures as test expressions in its
clauses, but it keeps evaluating such expressions upon every char
read from the port.

Instead the test expression creating the matcher procedure should be
evaluated only once, and that procedure should be stored in a binding,
and referenced only as the test expression in the clauses themselves.

Since (ice-9 expect) still uses un-hygienic defmacro, which cannot
work together with the hygienic macro system, it was necessary
to completely rewrite the expect implementation hygienically.